ingesting/requirements.txt
fastapi==0.116.1
loguru==0.7.0
uvicorn==0.29.0
pinecone==5.4.0
google-cloud-storage==2.18.2
pillow==10.4.0
python-multipart==0.0.9
requests==2.32.3